Force int32 in bimap; TODO fix packages under syntax

+++ bimap/bimap.go
@@ -6,19 +6,19 @@ package bimap
 // Speical constants that may be returned from certain methods that
 // access a Map.
 const (
-	NoInt int = -1
+	NoInt int32 = -1
 )
 
 // Map is a bi-directional mapping between strings and integers. The
 // integers forms a dense range starting a 0.
 type Map struct {
-	strToInt map[string]int
+	strToInt map[string]int32
 	intToStr []string
 }
 
 // New creates an empty Map
 func New() *Map {
-	return &Map{make(map[string]int), make([]string, 0, 1024)}
+	return &Map{make(map[string]int32), make([]string, 0, 1024)}
 }
 
 // FromSlice creates an empty Map from the given slice
@@ -32,13 +32,13 @@ func FromSlice(s []string) *Map {
 
 // Add adds the given string into the map and returns its id. The
 // string being added should not be empty. This is not thread safe.
-func (m *Map) Add(s string) int {
+func (m *Map) Add(s string) int32 {
 	if len(s) == 0 {
 		panic("trying to add an empty string")
 	}
 	i, ok := m.strToInt[s]
 	if !ok {
-		i = len(m.intToStr)
+		i = int32(len(m.intToStr))
 		m.strToInt[s] = i
 		m.intToStr = append(m.intToStr, s)
 	}
@@ -47,7 +47,7 @@ func (m *Map) Add(s string) int {
 
 // FindByString finds the id or returns NoInt if the string is not in
 // the map.
-func (m *Map) FindByString(s string) int {
+func (m *Map) FindByString(s string) int32 {
 	i, ok := m.strToInt[s]
 	if ok {
 		return i
@@ -58,8 +58,8 @@ func (m *Map) FindByString(s string) int {
 // FindByInt finds the string corresponding to the given integral
 // id. Returns the string if the id is in the map; or an empty string
 // if it is not.
-func (m *Map) FindByInt(i int) string {
-	if 0 <= i && i < len(m.intToStr) {
+func (m *Map) FindByInt(i int32) string {
+	if 0 <= i && i < int32(len(m.intToStr)) {
 		return m.intToStr[i]
 	}
 	return ""
@@ -67,7 +67,7 @@ func (m *Map) FindByInt(i int) string {
 
 // AppendByString translates a slice of string and appends the result
 // to the given slice. New words are also added to the Map.
-func (m *Map) AppendByString(strs []string, ints *[]int) {
+func (m *Map) AppendByString(strs []string, ints *[]int32) {
 	for _, s := range strs {
 		*ints = append(*ints, m.Add(s))
 	}
@@ -75,8 +75,8 @@ func (m *Map) AppendByString(strs []string, ints *[]int) {
 
 // TranslateByString translates a slice of string into a slice of
 // integers. New words are also added to the Map.
-func (m *Map) TranslateByString(strs []string) []int {
-	ints := make([]int, len(strs))
+func (m *Map) TranslateByString(strs []string) []int32 {
+	ints := make([]int32, len(strs))
 	for i, s := range strs {
 		ints[i] = m.Add(s)
 	}
@@ -85,7 +85,7 @@ func (m *Map) TranslateByString(strs []string) []int {
 
 // AppendByInt translates a slice of integers and appends the result
 // to the given slice.
-func (m *Map) AppendByInt(ints []int, strs *[]string) {
+func (m *Map) AppendByInt(ints []int32, strs *[]string) {
 	for _, id := range ints {
 		*strs = append(*strs, m.FindByInt(id))
 	}
@@ -93,7 +93,7 @@ func (m *Map) AppendByInt(ints []int, strs *[]string) {
 
 // TranslateByInt translates a slice of integers into a slice of
 // strings.
-func (m *Map) TranslateByInt(ints []int) []string {
+func (m *Map) TranslateByInt(ints []int32) []string {
 	strs := make([]string, len(ints))
 	for i, id := range ints {
 		strs[i] = m.FindByInt(id)
@@ -103,6 +103,6 @@ func (m *Map) TranslateByInt(ints []int) []string {
 
 // Size returns the size of the map, which is also the next id to be
 // assigned.
-func (m *Map) Size() int {
-	return len(m.intToStr)
+func (m *Map) Size() int32 {
+	return int32(len(m.intToStr))
 }
