+++ bimap/bimap_test.go
@@ -11,15 +11,15 @@ func TestMap(t *testing.T) {
 		t.Errorf("expected empty map; got size %d\n", size)
 	}
 	for i, s := range strs {
-		if id := m.Add(s); id != i {
+		if id := m.Add(s); id != int32(i) {
 			t.Errorf("expected %d; got %d\n", i, id)
 		}
 	}
 	for i, s := range strs {
-		if id := m.FindByString(s); id != i {
+		if id := m.FindByString(s); id != int32(i) {
 			t.Errorf("expected %d; got %d\n", i, id)
 		}
-		if ss := m.FindByInt(i); ss != s {
+		if ss := m.FindByInt(int32(i)); ss != s {
 			t.Errorf("expected %q; got %q\n", s, ss)
 		}
 	}
@@ -35,33 +35,33 @@ func TestMap(t *testing.T) {
 
 	m = FromSlice(strs)
 	for i, s := range strs {
-		if id := m.Add(s); id != i {
+		if id := m.Add(s); id != int32(i) {
 			t.Errorf("expected %d; got %d\n", i, id)
 		}
 	}
 
-	var intArray [3]int
+	var intArray [3]int32
 	intSlice := intArray[:0]
 	m.AppendByString(strs, &intSlice)
-	if intArray != [...]int{0, 1, 2} {
+	if intArray != [...]int32{0, 1, 2} {
 		t.Errorf("expected [0, 1, 2]; got %v\n", intArray)
 	}
 
 	if n := copy(intArray[:], m.TranslateByString(strs)); n != 3 {
 		t.Errorf("expected %d elements copied; got %d\n", 3, n)
 	}
-	if intArray != [...]int{0, 1, 2} {
+	if intArray != [...]int32{0, 1, 2} {
 		t.Errorf("expected [0, 1, 2]; got %v\n", intArray)
 	}
 
 	var strArray [3]string
 	strSlice := strArray[:0]
-	m.AppendByInt([]int{0, 1, 2}, &strSlice)
+	m.AppendByInt([]int32{0, 1, 2}, &strSlice)
 	if strArray != [...]string{"a", "b", "c"} {
 		t.Errorf("expected [a, b, c]; got %v\n", strArray)
 	}
 
-	if n := copy(strArray[:], m.TranslateByInt([]int{0, 1, 2})); n != 3 {
+	if n := copy(strArray[:], m.TranslateByInt([]int32{0, 1, 2})); n != 3 {
 		t.Errorf("expected %d element copied; got %d\n", 3, n)
 	}
 	if strArray != [...]string{"a", "b", "c"} {
